requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
Django~=4.2.19
django-cors-headers==4.6.0
django-restframework==0.0.1
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
executing==2.1.0
icecream==2.1.3
immutabledict==4.2.1
inflection==0.5.1
iniconfig==2.0.0
openpyxl==3.1.5
ortools==9.11.4210
packaging==24.2
pluggy==1.5.0
protobuf==5.26.1
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1

DateTime~=5.5
psycopg2~=2.9.10
pandas~=2.2.3